requirements.txt
discord.py
giphy_client
random
os
dotenv